Move files to new tasks folder and use getAllSectionsList()
