+++ src/modules/shared/pages/innovation/tasks/wizard-task-new/steps/section-step.component.html
@@ -4,7 +4,7 @@
     <div class="nhsuk-grid-column-two-thirds">
 
       <form [formGroup]="form" (ngSubmit)="onNextStep()">
-        <theme-form-radio-group controlName="section" [items]="sectionItems" [pageUniqueField]="false"></theme-form-radio-group>
+        <theme-form-radio-group controlName="section" [items]="data.sections" [pageUniqueField]="false"></theme-form-radio-group>
         <button type="submit" class="nhsuk-button nhsuk-u-margin-top-3">Continue</button>
       </form>
