+++ src/modules/shared/pages/innovation/tasks/wizard-task-new/steps/section-step.types.ts
@@ -1,9 +1,8 @@
 import { InnovationSectionEnum } from '@modules/stores/innovation';
-import { InnovationSectionsListType } from '@modules/stores/innovation/innovation-record/ir-versions.types';
 
 
 export type SectionStepInputType = {
-  sections: InnovationSectionsListType,
+  sections: { value: string, label: string }[],
   selectedSection: null | InnovationSectionEnum
 };
