+++ src/modules/shared/pages/innovation/tasks/wizard-task-new/task-new.component.ts
@@ -5,8 +5,7 @@ import { WizardModel, WizardStepModel } from '@app/base/models';
 import { MappedObjectType, WizardStepEventType } from '@app/base/types';
 import { InnovationsService } from '@modules/shared/services/innovations.service';
 import { InnovationSectionEnum } from '@modules/stores/innovation';
-import { getInnovationRecordConfig } from '@modules/stores/innovation/innovation-record/ir-versions.config';
-import { InnovationSectionsListType } from '@modules/stores/innovation/innovation-record/ir-versions.types';
+import { getAllSectionsList } from '@modules/stores/innovation/innovation-record/ir-versions.config';
 import { WizardTaskNewMessageStepComponent } from './steps/message-step.component';
 import { MessageStepInputType, MessageStepOutputType } from './steps/message-step.types';
 import { WizardTaskNewSectionStepComponent } from './steps/section-step.component';
@@ -25,7 +24,7 @@ export class PageInnovationTaskNewComponent extends CoreComponent implements OnI
   taskUrl: string;
   sectionUrl: string;
 
-  sections: InnovationSectionsListType;
+  sections: { value: string, label: string }[];
 
   wizard = new WizardModel<{
     sectionStep: { section: null | InnovationSectionEnum },
@@ -50,7 +49,7 @@ export class PageInnovationTaskNewComponent extends CoreComponent implements OnI
 
     this.sectionUrl = `/${this.baseUrl}/innovations/${this.innovationId}/record/sections/${this.sectionId}`;
 
-    this.sections = getInnovationRecordConfig();
+    this.sections = getAllSectionsList();
 
     this.wizard.data = {
       sectionStep: { section: null },
