+++ src/modules/shared/shared.module.ts
@@ -24,9 +24,9 @@ import { PageNotFoundComponent } from './pages/error/not-found.component';
 import { PageInnovationActionSectionInfoComponent } from './pages/innovation/actions/action-section-info.component';
 import { PageInnovationTaskToDoListComponent } from './pages/innovation/actions/task-to-do-list.component';
 import { PageTaskStatusListComponent } from './pages/innovation/actions/task-status-list.component';
-import { PageInnovationTaskNewComponent } from './pages/innovation/actions/wizard-task-new/task-new.component';
-import { WizardTaskNewSectionStepComponent } from './pages/innovation/actions/wizard-task-new/steps/section-step.component';
-import { WizardTaskNewMessageStepComponent } from './pages/innovation/actions/wizard-task-new/steps/message-step.component';
+import { PageInnovationTaskNewComponent } from './pages/innovation/tasks/wizard-task-new/task-new.component';
+import { WizardTaskNewSectionStepComponent } from './pages/innovation/tasks/wizard-task-new/steps/section-step.component';
+import { WizardTaskNewMessageStepComponent } from './pages/innovation/tasks/wizard-task-new/steps/message-step.component';
 import { PageInnovationActivityLogComponent } from './pages/innovation/activity-log/innovation-activity-log.component';
 import { PageInnovationAssessmentOverviewComponent } from './pages/innovation/assessment/assessment-overview.component';
 import { PageInnovationDataSharingAndSupportComponent } from './pages/innovation/data-sharing-and-support/data-sharing-and-support.component';
